Move Editor specific code out of production files, fixing build error.
