+++ Assets/Objects/Platforms/_shared/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 027a71ebafa1a7adeae3404f0a2e589a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
