+++ Assets/Objects/Platforms/_shared/Editor/MoveToWaypointsEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d64960c6d0ca1cc15b6e325509629f3b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
